Fixed Error with multiple track types
The error was that the system searches for the last element of an array of tracks and the hierarchy was sorting it in alphabetical order so therefore the system would not move horizontal tracks after a vertical track is spawn. The solution to this was to set beginplay on each track type bp to have its own unique tag that it shares with grid manager. Now when grid manager searches for tags it only searches the last element of the array of the most recent track type to be spawned.